Refactor: make package and classify
