+++ src/main/java/racingcar/Application.java
@@ -1,5 +1,9 @@
 package racingcar;
 
+import racingcar.game.RacingGame;
+import racingcar.view.RacingInputView;
+import racingcar.view.RacingOutputView;
+
 public class Application {
     public static void main(String[] args) {
         RacingInputView inputView = new RacingInputView();
