+++ src/test/java/racingcar/CarTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.Test;
+import racingcar.car.Car;
 
 class CarTest {
     private static final int MOVING_FORWARD = 4;
