+++ src/test/java/racingcar/CompeteCarsTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
 import org.junit.jupiter.api.Test;
+import racingcar.car.Car;
+import racingcar.car.CompeteCars;
 
 class CompeteCarsTest {
     @Test
