+++ src/test/java/racingcar/RacingOutputViewTest.java
@@ -9,6 +9,7 @@ import java.util.List;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import racingcar.view.RacingOutputView;
 
 class RacingOutputViewTest {
     private final RacingOutputView outputView = new RacingOutputView();
